pyproject.toml
[project]
name = "namesilo-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.24.0",
    "python-dotenv>=1.2.1",
    "xmltodict>=1.0.2",
]
